Added some "skeleton" code. (Empty files.)
This is most of the required FLTK source code files laid out.  It should provide a good starting point for figuring out how exactly to implement this thing.  It's mostly based on a quick diagram of how I expect the application to look when finished.

There are two main groups: windows and widgets.  Windows will extend FL_Window (while subwindows are used like widgets in the main windows).  Currently the subwindows directory contains FL_GL_Window objects.  The wdigets directory will contain easy to add widgets to the windows.

The primary purpose of the number of classes is to carefully encapsulate and isolate the functionality of the program.  This should make making changes and modifications easier in the future.